remove active selection on double click and persist selections
